Delete order_balance_filter
