+++ crates/solver/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod interactions;
 pub mod liquidity;
 pub mod liquidity_collector;
-pub mod order_balance_filter;
 pub mod settlement;
 pub mod solver;
